Combined dev and prod logger
